input: fix the wrong mouse wheel command often being used with touchpads
Options::wheelToMouseCommand silently truncated the floating point values to
integers, making it choose the wrong command most of the time

+++ src/options.cpp
@@ -836,7 +836,7 @@ bool Options::condensedTitle() const
     return condensed_title;
 }
 
-Options::MouseCommand Options::wheelToMouseCommand(MouseWheelCommand com, int delta) const
+Options::MouseCommand Options::wheelToMouseCommand(MouseWheelCommand com, qreal delta) const
 {
     switch (com) {
     case MouseWheelRaiseLower:
